feat: ts migration
