+++ package.json
@@ -1,15 +1,15 @@
 {
-  "name": "basic-nodejs-mitso",
+  "name": "rest-service-ts",
   "version": "1.0.0",
   "repository": {
     "type": "git",
     "url": ""
   },
   "type": "commonjs",
-  "description": "nodejs-mitso",
-  "main": "src/server.js",
+  "description": "REST service",
+  "main": "src/server.ts",
   "scripts": {
-    "start": "nodemon src/server.js",
+    "start": "nodemon src/server.ts",
     "lint": "eslint --ext .js,.ts --fix src/"
   },
   "keywords": [
@@ -43,20 +43,39 @@
     "cross-env": "7.0.3",
     "dotenv": "9.0.0",
     "express": "4.17.1",
+    "express-async-handler": "^1.1.4",
     "http-status-codes": "^2.1.4",
     "lint": "^0.7.0",
     "uuid": "8.3.2"
   },
   "devDependencies": {
+    "@types/dotenv": "^8.2.0",
+    "@types/express": "^4.17.12",
+    "@types/node": "^15.6.1",
+    "@types/swagger-ui-express": "^4.1.2",
+    "@types/uuid": "^8.3.0",
+    "@types/yamljs": "^0.2.31",
+    "@typescript-eslint/eslint-plugin": "^4.25.0",
+    "@typescript-eslint/parser": "^4.25.0",
+    "chai": "4.3.4",
     "debug": "4.3.1",
-    "eslint": "7.2.0",
-    "eslint-config-airbnb-base": "14.2.1",
+    "dirty-chai": "2.0.1",
+    "eslint": "^7.27.0",
+    "eslint-config-airbnb-base": "^14.2.1",
+    "eslint-config-airbnb-typescript": "^12.3.1",
+    "eslint-config-airbnb": "^18.2.1",
     "eslint-config-prettier": "8.3.0",
-    "eslint-plugin-import": "2.22.1",
+    "eslint-plugin-import": "^2.23.3",
     "eslint-plugin-node": "11.1.0",
     "eslint-plugin-prettier": "3.4.0",
+    "husky": "4.3.8",
+    "jest": "26.6.3",
+    "jsdoc": "^3.6.7",
     "lint-staged": "10.5.4",
     "nodemon": "2.0.7",
-    "prettier": "2.2.1"
+    "prettier": "2.2.1",
+    "supertest": "6.1.3",
+    "ts-node": "^10.0.0",
+    "typescript": "^4.3.2"
   }
 }
