+++ src/server.ts
@@ -1,0 +1,6 @@
+import config from './common/config';
+import app from  './app';
+
+const { PORT } = config;
+
+app.listen(PORT, () => process.stdout.write(`App is running on http://localhost:${PORT}`));
